pyproject.toml
[project]
name = "reddit-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.14.0",
    "praw>=7.8.1",
    "python-dotenv>=1.1.0",
]
